+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - HOWTO for creating device extension
+- Function to get a copy of caqtus serialization converter
 
 ### Changed